% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ending.R
\name{date_ending}
\alias{date_ending}
\title{This function produces the correct ordinal indicator for the date}
\usage{
date_ending(date)
}
\arguments{
\item{date}{A number}
}
\description{
This function produces the correct ordinal indicator for the date
}
\examples{
date_ending(1) # returns "st"
day(3) # returns "rd"

}
