% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{This function outputs the 'time before first proven reoffence' graph}
\usage{
forest_plot(data, subset, title, x_axis, y_axis)
}
\arguments{
\item{data}{The standard result dataset}

\item{subset}{Generic text size in graph}

\item{title}{Text size of confidence intervals}

\item{x_axis}{Text size of y-axis label}

\item{y_axis}{Text size of legend text}
}
\description{
This function outputs the 'time before first proven reoffence' graph
}
