% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{Formats numbers with a comma delimiter and values of less than 10 are represented as words}
\usage{
format_num(number, digits = 0, word = FALSE, capital = FALSE,
  allow_neg = FALSE, p_value = FALSE)
}
\arguments{
\item{number}{A number}

\item{digits}{A number representing number of decimal places}

\item{word}{A logical word showing whether the number is to be repsented as a word}

\item{capital}{A logical word showing whether the start of the word should be capatalised}

\item{allow_neg}{A logical word showing whether the number is allowed to be displayed as negative}

\item{p_value}{A logical word showing whether the number is a p-value}
}
\description{
Formats numbers with a comma delimiter and values of less than 10 are represented as words
}
\examples{
format_num("1000") # returns "1,000"
format_num("1", word = TRUE) # returns 'one'
}
