% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_or_lower_CI.R
\name{higher_or_lower_CI}
\alias{higher_or_lower_CI}
\title{This function produces a sentence that decribes a confidence interval in terms of people, offences or days}
\usage{
higher_or_lower_CI(no1, no2, type, gender)
}
\arguments{
\item{no1}{A number}

\item{no2}{A number}

\item{type}{A word}

\item{gender}{A word}
}
\description{
This function produces a sentence that decribes a confidence interval in terms of people, offences or days
}
\examples{
higher_or_lower_CI(1234,2345,"days",JDLpackage:::people_func(2,"male")) #returns higher by between 1,234 and 2,345 days. This is a statistically significant result.
higher_or_lower_CI(-1234,2345,"offences",JDLpackage:::people_func(2,"female")) #returns lower by as many as 1,234 offences, or higher by as many as 2,345 offences. More women would need to be analysed in order to determine the direction of this difference.
}
