% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon}
\alias{icon}
\title{Takes a p-value the difference, and outputs an icon that indicates whether it is an increase or
 decrease and whether it is significant or not in the number and frequency of reoffences}
\usage{
icon(pvalue, diff, stat)
}
\arguments{
\item{pvalue}{A number}

\item{diff}{A number}

\item{stat}{Shows what variable the icon is representing}
}
\description{
Takes a p-value the difference, and outputs an icon that indicates whether it is an increase or
 decrease and whether it is significant or not in the number and frequency of reoffences
}
\examples{
icon()


}
