% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralising_functions.R
\name{offence_func}
\alias{offence_func}
\title{This function outputs the correct singular or plural version of offence}
\usage{
offence_func(no)
}
\arguments{
\item{no}{A number}
}
\description{
This function outputs the correct singular or plural version of offence
}
\examples{
offence_func(1) #returns "offence"
offence_func(-3) #returns "offences"
}
