% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{percentage}
\alias{percentage}
\title{Takes two numbers and returns the percentage of the second number compared to the first one
If the rounded number equals 0, it will return with "<1\%"}
\usage{
percentage(no1, no2)
}
\arguments{
\item{no1}{A number}

\item{no2}{A number}
}
\description{
Takes two numbers and returns the percentage of the second number compared to the first one
If the rounded number equals 0, it will return with "<1\%"
}
\examples{
percentage(20,5) #returns "25\%"
percentage(100,1) #returns "1\%"
}
