% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralising_functions.R
\name{plural_analysis}
\alias{plural_analysis}
\title{Formats the word 'analysis' as being plural and have a capital if needed}
\usage{
plural_analysis(number, capital = FALSE)
}
\arguments{
\item{number}{A number}

\item{capital}{A lagical word showing whether the start of the word should be capatalised}
}
\description{
Formats the word 'analysis' as being plural and have a capital if needed
}
\examples{
format_num("2") # returns "analyses"
format_num("1", capital = TRUE) # returns "Analysis"


}
