% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reoff_freq_graph.R
\name{reoff_freq_graph}
\alias{reoff_freq_graph}
\title{This function outputs the 'one-year proven re-offending rate' graph}
\usage{
reoff_freq_graph(data, max_x_val, x_ticks, text_size = 11, ci_text_size = 5,
  y_title_size = 11, legend_text_size = 12, comp_CI_x_position = 0,
  treat_CI_x_position = 0, comp_CI_y_position = 0,
  treat_CI_y_position = 0, bar_width = 0.9)
}
\arguments{
\item{data}{The standard result dataset}

\item{text_size}{Generic text size in graph}

\item{ci_text_size}{Text size of confidence intervals}

\item{y_title_size}{Text size of y-axis label}

\item{legend_text_size}{Text size of legend text}

\item{comp_CI_x_position}{Horizontal position of the comparison group CI label with respect to the x-axis}

\item{comp_CI_y_position}{Vertical position of the comparison group CI label with a unit being the space between the center of the bars}

\item{bar_width}{The width of the bars}

\item{reoff_CI_x_position}{Horizontal position of the treatment group CI label with respect to the x-axis}

\item{reoff_CI_y_position}{Vertical position of the treatment group CI label with a unit being the space between the center of the bars}
}
\description{
This function outputs the 'one-year proven re-offending rate' graph
}
