% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorter_or_longer_CI.R
\name{shorter_or_longer_CI}
\alias{shorter_or_longer_CI}
\title{This function produces a sentence that decribes a confidence interval in terms of people, offences or days}
\usage{
shorter_or_longer_CI(no1, no2, type, gender)
}
\arguments{
\item{no1}{A number}

\item{no2}{A number}

\item{type}{A word}

\item{gender}{A word}
}
\description{
This function produces a sentence that decribes a confidence interval in terms of people, offences or days
}
\examples{
shorter_or_longer_CI(1234,2345,days,male) #returns longer by between 1,234 and 2,345 days. This is a statistically significant result.
shorter_or_longer_CI(-1234,2345,offences,female) #returns shorter by as many as 1,234 offences, or longer by as many as 2,345 offences. More women would need to be analysed in order to determine the direction of this difference.
}
