% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_or_non_sig.R
\name{sig_or_non_sig}
\alias{sig_or_non_sig}
\title{This function outputs either more or less depending on the difference between the two numbers}
\usage{
sig_or_non_sig(p_value)
}
\arguments{
\item{p_value}{A number}
}
\description{
This function outputs either more or less depending on the difference between the two numbers
}
\examples{
more_or_less(0.01) #returns "**Significant difference between groups**"
more_or_less(0.5) #returns "**Non-significant difference between groups**"
}
