% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_table_line.R
\name{sig_table_line}
\alias{sig_table_line}
\title{This function outputs either more or less depending on the difference between the two numbers}
\usage{
sig_table_line(p_value, bracket, class, ...)
}
\arguments{
\item{p_value}{The p-vlaue realated to the variables being showned on that table line}

\item{bracket}{Type of HTML table bracket to be applied to the line}

\item{class}{CSS class to apply to line}

\item{...}{The test to be put in each cell in the line}
}
\description{
This function outputs either more or less depending on the difference between the two numbers
}
\examples{
more_or_less(0.01) #returns "**Significant difference between groups**"
more_or_less(0.5) #returns "**Non-significant difference between groups**"

}
