% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/was_were.R
\name{was_were}
\alias{was_were}
\title{Produces sigular or plural version of was}
\usage{
was_were(number, capital = FALSE)
}
\arguments{
\item{number}{A number}

\item{capital}{A lagical word showing whether the start of the word should be capatalised}
}
\description{
Produces sigular or plural version of was
}
\examples{
format_num("2") # returns "were"
format_num("1", capital = TRUE) # returns "Was"
}
